+++ dsp/dsp.cpp
@@ -20,7 +20,7 @@ void dsp::DSP::_AllocateOutputPointers(const size_t numChannels) {
   if (this->mOutputPointers != nullptr)
     throw std::runtime_error(
         "Tried to re-allocate over non-null mOutputPointers");
-  this->mOutputPointers = new float *[numChannels];
+  this->mOutputPointers = new double *[numChannels];
   if (this->mOutputPointers == nullptr)
     throw std::runtime_error("Failed to allocate pointer to output buffer!\n");
   this->mOutputPointersSize = numChannels;
@@ -36,7 +36,7 @@ void dsp::DSP::_DeallocateOutputPointers() {
   this->mOutputPointersSize = 0;
 }
 
-float **dsp::DSP::_GetPointers() {
+double **dsp::DSP::_GetPointers() {
   for (auto c = 0; c < this->_GetNumChannels(); c++)
     this->mOutputPointers[c] = this->mOutputs[c].data();
   return this->mOutputPointers;
@@ -91,7 +91,7 @@ void dsp::History::_RewindHistory() {
   this->mHistoryIndex = this->mHistoryRequired;
 }
 
-void dsp::History::_UpdateHistory(float **inputs, const size_t numChannels,
+void dsp::History::_UpdateHistory(double **inputs, const size_t numChannels,
                                   const size_t numFrames) {
   this->_EnsureHistorySize(numFrames);
   if (numChannels < 1)
